chore: retirer ce que plus personne ne lit, et garder la révision honnête
`withdrawn` sortait du contrat de progression sans lecteur : la webapp n'en
avait aucune occurrence (son `row.withdrawn` est une ligne de timeline dérivée
de `month.withdrawnAmount`, autre chose), et le modèle iOS ne le décode pas —
Swift ignore les clés en trop, donc le retrait ne casse aucun client déployé.
`confirmed` porte déjà l'information : la somme y est soustraite.

Six membres de CurrentMonthStore avaient perdu tous leurs lecteurs à la refonte
de l'accueil. Leurs tests ne les appelaient plus : chacun réimplémentait le
filtre dans sa propre phase Act, donc ils passaient déjà sans le store. Seuls
les cas du report survivent, chez BudgetFormulas dont ils testent la fonction.

Enfin, la date de début décide de la fenêtre de contribution, donc du solde
confirmé — la déplacer changeait ce solde sans toucher un montant, et sans
faire bouger la révision qui le protège : un retrait calculé sur l'ancienne
fenêtre restait accepté. La liste de colonnes du déclencheur est la moitié
porteuse, `UPDATE OF` ne se déclenche que sur les colonnes citées.
