+++ frontend/e2e/tests/features/savings-goals-progress.spec.ts
@@ -69,7 +69,6 @@ const progress = {
   plannedCumulative: 1200,
   plannedProjection: 1200,
   confirmed: 900,
-  withdrawn: 0,
   achievementPercent: 30,
   monthsElapsed: 2,
   monthsRemaining: 14,
@@ -147,7 +146,6 @@ function progressFor(goalState: SavingsGoal): SavingsGoalProgress {
     plannedCumulative: 300,
     plannedProjection: 600,
     confirmed: 300,
-    withdrawn: 0,
     achievementPercent: hasTarget ? 10 : null,
     monthsElapsed: 1,
     monthsRemaining: hasDeadline ? 2 : null,
