+++ frontend/projects/webapp/src/app/feature/savings-goals/detail/components/goal-plan-simulator-toolbar.spec.ts
@@ -90,7 +90,6 @@ function makeProgress(
     plannedProjection: 400,
     confirmed: 0,
     initialAmount: 0,
-    withdrawn: 0,
     achievementPercent: 0,
     monthsElapsed: 1,
     monthsRemaining: 2,
