+++ frontend/projects/webapp/src/app/feature/savings-goals/detail/services/goal-plan-simulator-store.spec.ts
@@ -47,7 +47,6 @@ function makeProgress(
     plannedProjection: 400,
     confirmed: 0,
     initialAmount: 0,
-    withdrawn: 0,
     achievementPercent: 0,
     monthsElapsed: 1,
     monthsRemaining: 2,
