+++ frontend/projects/webapp/src/app/feature/savings-goals/services/savings-goals-store.spec.ts
@@ -50,7 +50,6 @@ function makeProgress(
     plannedProjection: 1200,
     confirmed: 900,
     initialAmount: 0,
-    withdrawn: 0,
     achievementPercent: 30,
     monthsElapsed: 3,
     monthsRemaining: 12,
